maintain one connection to Redis by separating connecting to Redis in another module; close connection when process ends
